refactor(sdk-go): split reasoner registration methods
Move RegisterReasoner out of agent.go into agent_register.go as a pure file split to reduce file size and keep behavior unchanged.

Made-with: Cursor
